OSTJ-119: Refaktorointia
