+++ osoitekoostepalvelu/src/main/java/fi/vm/sade/osoitepalvelu/kooste/service/search/MessageService.java
@@ -1,0 +1,7 @@
+package fi.vm.sade.osoitepalvelu.kooste.service.search;
+
+import java.util.Locale;
+
+public interface MessageService {
+    String getMessage(String key, Locale locale, Object... params);
+}
