+++ osoitekoostepalvelu/src/main/java/fi/vm/sade/osoitepalvelu/kooste/service/search/MessageServiceImpl.java
@@ -1,0 +1,24 @@
+package fi.vm.sade.osoitepalvelu.kooste.service.search;
+
+import fi.vm.sade.osoitepalvelu.kooste.service.AbstractService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
+import org.springframework.context.NoSuchMessageException;
+import org.springframework.stereotype.Service;
+
+import java.util.Locale;
+
+@Service
+public class MessageServiceImpl extends AbstractService implements MessageService {
+
+    @Autowired
+    private MessageSource messageSource;
+
+    public String getMessage(String key, Locale locale, Object... params) {
+        try {
+            return messageSource.getMessage(key, params, locale);
+        } catch (NoSuchMessageException e) {
+            return messageSource.getMessage(key, params, DEFAULT_LOCALE);
+        }
+    }
+}
