+++ osoitekoostepalvelu/src/main/java/fi/vm/sade/osoitepalvelu/kooste/service/search/detailcopier/KoskiYhdyshenkiloCopier.java
@@ -1,0 +1,29 @@
+package fi.vm.sade.osoitepalvelu.kooste.service.search.detailcopier;
+
+import fi.vm.sade.osoitepalvelu.kooste.common.util.CollectionHelper;
+import fi.vm.sade.osoitepalvelu.kooste.route.dto.OrganisaatioDetailsDto;
+import fi.vm.sade.osoitepalvelu.kooste.route.dto.OrganisaatioYhteystietoElementtiDto;
+import fi.vm.sade.osoitepalvelu.kooste.route.dto.helpers.OrganisaatioYksityiskohtainenYhteystietoByKoskiYhdyshenkiloPredicate;
+import fi.vm.sade.osoitepalvelu.kooste.service.search.dto.SearchResultRowDto;
+
+import java.util.Iterator;
+import java.util.Locale;
+
+import static fi.vm.sade.osoitepalvelu.kooste.service.AbstractService.DEFAULT_LOCALE;
+
+public class KoskiYhdyshenkiloCopier implements DetailCopier {
+    @Override
+    public boolean isMissing(SearchResultRowDto from) { return from.getKoskiYhdyshenkilo() == null; }
+
+    @Override
+    public void copy(OrganisaatioDetailsDto from, SearchResultRowDto to, Locale locale) {
+        Iterator<OrganisaatioYhteystietoElementtiDto> yhteystietoArvos =
+                CollectionHelper.filter(from.getYhteystietoArvos(),
+                        new OrganisaatioYksityiskohtainenYhteystietoByKoskiYhdyshenkiloPredicate(locale),
+                        new OrganisaatioYksityiskohtainenYhteystietoByKoskiYhdyshenkiloPredicate(DEFAULT_LOCALE))
+                        .iterator();
+        if(yhteystietoArvos.hasNext()) {
+            to.setKoskiYhdyshenkilo(yhteystietoArvos.next().getArvo());
+        }
+    }
+}
