+++ osoitekoostepalvelu/src/main/java/fi/vm/sade/osoitepalvelu/kooste/service/search/detailcopier/OrganisaatiotunnisteCopier.java
@@ -1,0 +1,18 @@
+package fi.vm.sade.osoitepalvelu.kooste.service.search.detailcopier;
+
+import fi.vm.sade.osoitepalvelu.kooste.route.dto.OrganisaatioDetailsDto;
+import fi.vm.sade.osoitepalvelu.kooste.service.search.dto.SearchResultRowDto;
+
+import java.util.Locale;
+
+public class OrganisaatiotunnisteCopier implements DetailCopier {
+    @Override
+    public boolean isMissing(SearchResultRowDto from) {
+        return from.getOppilaitosKoodi() == null;
+    }
+
+    @Override
+    public void copy(OrganisaatioDetailsDto from, SearchResultRowDto to, Locale locale) {
+        to.setOppilaitosKoodi(from.getOppilaitosKoodi());
+    }
+}
